Refactor ChineseIME. Now it acts as a backend. Migrate yaml configuration GUI tools to settings. Now the implementations of Rime engine and settings GUI are unbound.
