+++ java/src/org/futo/inputmethod/engine/general/ChineseIME.kt
@@ -1,21 +1,13 @@
 package org.futo.inputmethod.engine.general
 
-import android.content.Context
 import android.os.SystemClock
 import android.util.Log
 import android.view.KeyCharacterMap
 import android.view.KeyEvent
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalResources
-import androidx.compose.ui.res.stringResource
-import androidx.datastore.preferences.core.stringPreferencesKey
 import icu.astronot233.rime.DeployStage
 import icu.astronot233.rime.Rime
 import icu.astronot233.rime.RimeMessage
-import icu.astronot233.rime.RimeSchema
 import icu.astronot233.rime.SyncStage
 import icu.astronot233.rime.X11Keys.XK_BackSpace
 import icu.astronot233.rime.X11Keys.XK_Linefeed
@@ -30,14 +22,12 @@ import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.sync.Semaphore
 import org.futo.inputmethod.engine.ExpandableSuggestionBarConfiguration
 import org.futo.inputmethod.engine.GlobalIMEMessage
 import org.futo.inputmethod.engine.IMEHelper
 import org.futo.inputmethod.engine.IMEInterface
 import org.futo.inputmethod.engine.IMEMessage
 import org.futo.inputmethod.event.Event
-import org.futo.inputmethod.latin.R
 import org.futo.inputmethod.latin.Subtypes.switchToNextLanguage
 import org.futo.inputmethod.latin.SuggestedWords
 import org.futo.inputmethod.latin.SuggestedWords.SuggestedWordInfo
@@ -46,176 +36,16 @@ import org.futo.inputmethod.latin.common.Constants
 import org.futo.inputmethod.latin.common.InputPointers
 import org.futo.inputmethod.latin.settings.Settings
 import org.futo.inputmethod.latin.suggestions.SuggestionStripViewAccessor
-import org.futo.inputmethod.latin.uix.FileKind
 import org.futo.inputmethod.latin.uix.FloatingPreEdit
 import org.futo.inputmethod.latin.uix.PreEditListener
-import org.futo.inputmethod.latin.uix.SettingsKey
-import org.futo.inputmethod.latin.uix.actions.throwIfDebug
-import org.futo.inputmethod.latin.uix.getSetting
-import org.futo.inputmethod.latin.uix.namePreferenceKeyFor
-import org.futo.inputmethod.latin.uix.preferenceKeyFor
-import org.futo.inputmethod.latin.uix.setSetting
-import org.futo.inputmethod.latin.uix.settings.CollapsibleSection
-import org.futo.inputmethod.latin.uix.settings.DropDownPickerSettingItem
-import org.futo.inputmethod.latin.uix.settings.SettingToggleRaw
-import org.futo.inputmethod.latin.uix.settings.UserSetting
-import org.futo.inputmethod.latin.uix.settings.UserSettingsMenu
-import org.futo.inputmethod.latin.uix.settings.useDataStore
-import org.futo.inputmethod.latin.utils.ZipFileHelper
-import org.futo.inputmethod.latin.utils.toEnumOrNull
+import org.futo.inputmethod.latin.uix.settings.pages.langspecific.zh.RimeSettings
 import org.futo.inputmethod.v2keyboard.KeyboardLayoutSetV2
-import java.io.File
-import java.util.Locale
 import kotlin.math.max
 import kotlin.math.min
 
-object ChineseIMESettings {
-    enum class ChineseSimplificationMode(val stringResource: Int) {
-        // These must not be renamed, or the existing setting value for users will break
-        ByLanguage(R.string.chinese_setting_simplification_simplify_by_language_country),
-        Simplified(R.string.chinese_setting_simplification_force_simplified),
-        Traditional(R.string.chinese_setting_simplification_force_traditional);
-    }
 
-    enum class PinyinScheme(
-        val displayNameZH: String,
-        val displayNameEN: String,
-        val rimeId: String
-    ) {
-        // These must not be renamed, or the existing setting value for users will break
-        FullPinyin("全拼", "Full Pinyin", "luna_pinyin"),
-        Natural("自然碼雙拼", "Double Pinyin", "double_pinyin"),
-        ABC("智能ABC雙拼", "Double Pinyin - ABC", "double_pinyin_abc"),
-        FlyPY("小鶴雙拼", "Double Pinyin - flyPY", "double_pinyin_flypy"),
-        MS("微軟雙拼", "Double Pinyin - MS", "double_pinyin_mspy"),
-        PYJJ("拼音加加雙拼", "Double Pinyin - PYJJ", "double_pinyin_pyjj"),
-        ST("四通雙拼", "Double Pinyin - ST", "double_pinyin_st");
-    }
-
-    internal fun initialPatch(pattern: String, substitution: String, twoWay: Boolean)
-            = buildList {
-        add("derive/^${pattern}/${substitution}/")
-        if(twoWay) add("derive/^${substitution}/${pattern}/")
-    }
-
-    internal fun finalPatch(pattern: String, substitution: String, twoWay: Boolean)
-            = buildList {
-        add("derive/${pattern}$/${substitution}/")
-        if(twoWay) add("derive/${substitution}$/${pattern}/")
-    }
-
-    enum class FuzzyPinyinModes(val schemePatch: List<String>, val displayName: String) {
-        // These must not be renamed, or the existing setting value for users will break
-        Z_ZH(initialPatch("z", "zh", true), "z = zh"),
-        C_CH(initialPatch("c", "ch", true), "c = ch"),
-        S_SH(initialPatch("s", "sh", true), "s = sh"),
-        L_N(initialPatch("l", "n", true), "l = n"),
-        F_H(initialPatch("f", "h", true), "f = h"),
-        R_L(initialPatch("r", "l", false), "r = l"),
-        K_G(initialPatch("k", "g", false), "k = g"),
-
-        AN_ANG(finalPatch("an", "ang", true), "an = ang"),
-        EN_ENG(finalPatch("en", "eng", true), "en = eng"),
-        IN_ING(finalPatch("in", "ing", true), "in = ing"),
-        IAN_IANG(finalPatch("ian", "iang", true), "ian = iang"),
-        UAN_UANG(finalPatch("uan", "uang", true), "uan = uang");
-
-        companion object {
-            @JvmStatic
-            fun fromCommaString(str: String) =
-                str.split(',').mapNotNull { it.toEnumOrNull<FuzzyPinyinModes>() }.toSet()
-
-            @JvmStatic
-            fun toCommaString(value: Set<FuzzyPinyinModes>) =
-                value.joinToString(",") { it.name }
-        }
-    }
-
-    val SimplificationSetting = SettingsKey(
-        stringPreferencesKey("ChineseIME_simplification"),
-        ChineseSimplificationMode.ByLanguage.name
-    )
-
-    val PinyinSchemeSetting = SettingsKey(
-        stringPreferencesKey("ChineseIME_pinyin_scheme"),
-        PinyinScheme.FullPinyin.name
-    )
-
-    val FuzzyPinyinSetting = SettingsKey(
-        stringPreferencesKey("ChineseIME_fuzzy_pinyin"),
-        ""
-    )
-
-    var schemaList: List<RimeSchema> = emptyList()
-    val menu = UserSettingsMenu(
-        title = R.string.chinese_settings_title,
-        searchTags = R.string.chinese_setting_search_tags,
-        navPath = "ime/zh", registerNavPath = true,
-        settings = listOf(
-            UserSetting(R.string.chinese_setting_simplification) {
-                val resources = LocalResources.current
-                val (setting, setSetting) = useDataStore(SimplificationSetting)
-                DropDownPickerSettingItem<ChineseSimplificationMode>(
-                    stringResource(R.string.chinese_setting_simplification),
-                    ChineseSimplificationMode.entries,
-                    setting.toEnumOrNull<ChineseSimplificationMode>() ?: ChineseSimplificationMode.ByLanguage,
-                    { setSetting(it.name) },
-                    { resources.getString(it.stringResource) }
-                )
-            },
-
-            UserSetting(R.string.chinese_setting_pinyin_scheme) {
-                val context = LocalContext.current
-                val resources = LocalResources.current
-                val (setting, setSetting) = useDataStore(PinyinSchemeSetting)
-
-                val availableEntries = remember {
-                    val sharedDir = ChineseIME.getShared(context)
-                    val files = (sharedDir.listFiles() ?: emptyArray()).map { it.name.split('.').first() }.toSet()
-                    PinyinScheme.entries.filter {
-                        files.contains(it.rimeId) || it == PinyinScheme.FullPinyin
-                    }
-                }
-                DropDownPickerSettingItem<PinyinScheme>(
-                    stringResource(R.string.chinese_setting_pinyin_scheme),
-                    availableEntries,
-                    setting.toEnumOrNull<PinyinScheme>() ?: PinyinScheme.FullPinyin,
-                    { setSetting(it.name) },
-                    { when(resources.configuration.locale.language) {
-                        "zh" -> it.displayNameZH
-                        else -> it.displayNameEN
-                    } }
-                )
-            },
-
-            UserSetting(R.string.chinese_setting_fuzzy_pinyin) {
-                val (setting, setSetting) = useDataStore(FuzzyPinyinSetting)
-                val set = remember(setting) { FuzzyPinyinModes.fromCommaString(setting) }
-
-                CollapsibleSection(stringResource(R.string.chinese_setting_fuzzy_pinyin)) {
-                    FuzzyPinyinModes.entries.forEach { mode ->
-                        SettingToggleRaw(mode.displayName, set.contains(mode), { enabled ->
-                            val newSet = if (enabled) {
-                                set + setOf(mode)
-                            } else {
-                                set - setOf(mode)
-                            }
-
-                            setSetting(FuzzyPinyinModes.toCommaString(newSet))
-                        })
-                    }
-                }
-
-            },
-
-        )
-    )
-}
-
-class InputKeeper() {
+internal class InputKeeper {
     var text = ""
-        private set
-
     fun clear() {
         text = ""
     }
@@ -236,7 +66,8 @@ class InputKeeper() {
 class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAccessor, PreEditListener {
     private val TAG = "ChineseIME (rime)"
 
-    private val rime: Rime
+    private val rime: Rime =
+        Rime(RimeSettings.SharedDataDir, RimeSettings.UserDataDir, helper.context.packageName)
     private val connect get() = helper.getCurrentInputConnection()
     private val coroScope = CoroutineScope(Dispatchers.Main + SupervisorJob())
     private val expandableUiCfg = ExpandableSuggestionBarConfiguration(true, false)
@@ -245,12 +76,7 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
     private val rawInput = InputKeeper()
     private val rimeLoading = mutableStateOf(false)
 
-    override fun getLoadingState(): MutableState<Boolean>? = rimeLoading
-
-    private val currentTransformation get() = when(layoutHint) {
-        "stroke" -> StrokeTransformation
-        else -> emptyMap()
-    }
+    override fun getLoadingState() = rimeLoading
 
     private enum class EditingState(val active: Boolean) {
         NotEditing(false),
@@ -258,81 +84,6 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
         Finishing(false)
     }
     private var editingState = EditingState.NotEditing
-
-    companion object {
-        @JvmStatic
-        private val StrokeTransformation = mapOf(
-            'h' to '⼀',
-            's' to '⼁',
-            'p' to '⼃',
-            'n' to '⼂',
-            'z' to '⼄',
-        )
-
-        @JvmStatic
-        fun getRimeDir(context: Context) = context.getExternalFilesDir("rime")
-                ?: throw IllegalStateException("Failed to access ExternalFilesDir!")
-
-        @JvmStatic
-        fun getShared(context: Context) = File(getRimeDir(context), "shared")
-
-        @JvmStatic
-        fun getUser(context: Context) = File(getRimeDir(context), "user")
-
-        val PreviouslyExtractedDictionaryName = SettingsKey(
-            stringPreferencesKey("ChineseDictionaryExtractedValue"),
-            ""
-        )
-        var localPrevExtractedDictionaryName: String? = null
-
-        fun resetSharedFromResources(context: Context, scope: CoroutineScope): Boolean {
-            val pref = FileKind.Dictionary.preferenceKeyFor("zh")
-            val namePref = FileKind.Dictionary.namePreferenceKeyFor("zh")
-            val filePath = context.getSetting(pref, "")
-            val file = File(context.applicationContext.getExternalFilesDir(null), filePath)
-
-            val name = context.getSetting(namePref, "?")
-
-
-
-            if((localPrevExtractedDictionaryName ?: context.getSetting(PreviouslyExtractedDictionaryName)) == name) {
-                return false
-            }
-
-            try {
-                val shared = getShared(context)
-                shared.walkBottomUp()
-                    .fold(true) { res, it -> (it == shared || it.delete() || !it.exists()) && res }
-
-                // We are intentionally checking this only after deleting the shared directory
-                // This is because if the user deletes the dictionary file, we want to apply
-                // the deletion (TODO: It doesn't apply)
-                if(filePath.isEmpty()) return true
-                if(!file.exists()) return true
-
-                ZipFileHelper.extract(file, shared)
-
-                return true
-            } finally {
-                localPrevExtractedDictionaryName = name
-                scope.launch(Dispatchers.IO) {
-                    context.setSetting(PreviouslyExtractedDictionaryName.key, name)
-                }
-            }
-        }
-    }
-
-    init {
-        val shared = getShared(helper.context)
-        val user = getUser(helper.context)
-
-        if(!shared.exists() || !user.exists()) {
-            shared.mkdirs()
-            user.mkdirs()
-        }
-
-        rime = Rime(shared.path, user.path, helper.context.packageName)
-    }
 
     private fun handlePassByMessage(x11Code: Int, mask: Int) {
         when (x11Code) {
@@ -357,20 +108,13 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
         is RimeMessage.Deploy -> when (msg.value) {
             DeployStage.Unknown -> Log.e(TAG, "Deploy: Failed")
             DeployStage.Startup -> Log.i(TAG, "Deploy: Startup")
-            DeployStage.Success -> {
-                Log.i(TAG, "Deploy: Success")
-                coroScope.launch {
-                    ChineseIMESettings.schemaList = rime.getSchemata().also {
-                        Log.d(TAG, "Schemas: ${it.joinToString { it.toString() }} ")
-                    }
-                }
-            }
+            DeployStage.Success -> Log.i(TAG, "Deploy: Success")
             else -> {}
         }
         is RimeMessage.Sync -> when (msg.value) {
             SyncStage.Unknown -> Log.e(TAG, "Sync: Failed")
-//            SyncStage.Startup ->
-//            SyncStage.Success ->
+            SyncStage.Startup -> Log.i(TAG, "Sync: Startup")
+            SyncStage.Success -> Log.i(TAG, "Sync: Success")
             else -> {}
         }
         is RimeMessage.Commit -> {
@@ -390,7 +134,7 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
         if(ped.isEmpty() && notEditing) { rawInput.clear() }
 
         helper.updateUiInputState(ped.isEmpty() && notEditing)
-        helper.setPreedit(FloatingPreEdit.build(ped, this, currentTransformation, rawInput.text))
+        helper.setPreedit(FloatingPreEdit.build(ped, this, emptyMap(), rawInput.text))
     }.launchIn(coroScope)
 
     override fun onStartEdit() {
@@ -514,7 +258,6 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
                     IMEMessage.ReloadResources -> coroScope.launch {
                         rime.shutdown()
                         rime.startup(false)
-                        prevConfiguration = null
                     }
                     else -> {}
                 }
@@ -529,158 +272,10 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
         rime.shutdown()
     }
 
-    override fun onStartInput() {
-        if(resetSharedFromResources(helper.context, helper.lifecycleScope)) {
-            prevConfiguration = null
-            coroScope.launch { rime.deploy() }
-        }
-        updateConfig()
-    }
-
-    private fun isSimplifiedChinese(locale: Locale): Boolean {
-        if (locale.language != "zh") return false
-        val setting = helper.context.getSetting(ChineseIMESettings.SimplificationSetting).toEnumOrNull<ChineseIMESettings.ChineseSimplificationMode>()
-        return when(setting) {
-            ChineseIMESettings.ChineseSimplificationMode.Simplified -> true
-            ChineseIMESettings.ChineseSimplificationMode.Traditional -> false
-            else -> {
-                locale.script.takeIf { it.isNotEmpty() }?.let {
-                    return it.equals("Hans", ignoreCase = true)
-                }
-
-                return when (locale.country.uppercase()) {
-                    "TW", "HK", "MO" -> false
-                    else -> true
-                }
-            }
-        }
-    }
-
-
-
-    private data class Configuration(
-        val schema: String,
-        val learning: Boolean,
-        val simplification: Boolean,
-        val autocorrect: Boolean,
-        val fuzzyMode: Set<ChineseIMESettings.FuzzyPinyinModes>
-    )
-    private var prevConfiguration: Configuration? = null
-
-    private fun writeCustomizationFile(cfg: Configuration) {
-        val schema = cfg.schema
-        val file = File(getUser(helper.context), "${schema}.custom.yaml")
-        val content = buildString {
-            appendLine("patch:")
-
-            appendLine("    translator/enable_correction: ${cfg.autocorrect}")
-            appendLine("    translator/enable_user_dict: ${cfg.learning}")
-
-            appendLine("    switches/+:")
-            appendLine("        - name: futo_zh_simp")
-            appendLine("          reset: ${if(cfg.simplification) 1 else 0}")
-            appendLine("          states: [ 漢字, 汉字 ]")
-            appendLine("    simplifier/option_name: futo_zh_simp")
-
-            // Fuzzy pinyin only added to pinyin layouts
-            if(schema.startsWith("luna_pinyin")) {
-                appendLine("    speller/algebra:")
-                cfg.fuzzyMode.forEach {
-                    it.schemePatch.forEach {
-                        appendLine("        - $it")
-                    }
-                }
-            }
-        }
-
-        if(file.parentFile?.isDirectory == true) {
-            file.writeText(content)
-        }
-    }
-
-    private var configSemaphore = Semaphore(1)
-    private fun updateConfig() {
-        coroScope.launch {
-            if(editingState.active) return@launch
-            if(!configSemaphore.tryAcquire()) return@launch
-
-            try {
-                val settings = Settings.getInstance().current
-                val locale = settings.mLocale
-
-                val simplified = isSimplifiedChinese(locale)
-
-                val pinyinScheme = helper.context.getSetting(ChineseIMESettings.PinyinSchemeSetting)
-                    .toEnumOrNull<ChineseIMESettings.PinyinScheme>()
-                    ?: ChineseIMESettings.PinyinScheme.FullPinyin
-                val schema = when (layoutHint) {
-                    "qwerty" -> {
-                        when {
-                            pinyinScheme == ChineseIMESettings.PinyinScheme.FullPinyin -> when {
-                                simplified -> "luna_pinyin_simp"
-                                else -> "luna_pinyin"
-                            }
-
-                            else -> pinyinScheme.rimeId
-                        }
-                    }
-
-                    "stroke" -> {
-                        "stroke"
-                    }
-
-                    null -> {
-                        return@launch
-                    }
-
-                    else -> {
-                        throwIfDebug(IllegalStateException("Invalid layout hint '${layoutHint}'"))
-                        "luna_pinyin"
-                    }
-                }
-
-                val fuzzy = ChineseIMESettings.FuzzyPinyinModes.fromCommaString(
-                    helper.context.getSetting(ChineseIMESettings.FuzzyPinyinSetting)
-                )
-
-                val autocorrect = settings.mAutoCorrectionEnabledPerUserSettings
-                        || settings.isSuggestionsEnabledPerUserSettings
-
-                val learning = settings.isPersonalizationEnabled
-
-                val config = Configuration(schema, learning, simplified, autocorrect, fuzzy)
-                if (config != prevConfiguration) {
-                    rimeLoading.value = true
-                    writeCustomizationFile(config)
-                    rime.deploy()
-
-                    var result = false
-                    for(i in 0..10) {
-                        if(rime.selectSchema(config.schema)) {
-                            result = true
-                            break
-                        }
-                        Thread.sleep(300L)
-                    }
-
-                    if(!result) {
-                        Log.e(TAG, "Failed to select schema 3s after deployment!")
-                    }
-
-                    rime.setOption("simplification", simplified)
-                    rime.setOption("traditional", !simplified)
-                    prevConfiguration = config
-                }
-            } finally {
-                configSemaphore.release()
-                rimeLoading.value = false
-            }
-        }
-    }
+    override fun onStartInput() {}
 
     override fun onLayoutUpdated(layout: KeyboardLayoutSetV2) {
         layoutHint = layout.mainLayout.imeHint
-        if(helper.isImeActive(this)) updateConfig()
     }
 
     override fun onFinishInput() {
@@ -692,8 +287,7 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
         }
     }
 
-    override fun onFinishSlidingInput() {
-    }
+    override fun onFinishSlidingInput() {}
 
     private fun interruptInput(text: CharSequence? = null) {
         coroScope.launch {
@@ -792,16 +386,6 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
 
             Event.EVENT_TYPE_DOWN_UP_KEYEVENT -> {
                 interruptInput()
-                if (event.mX == KeyEvent.META_CTRL_ON) when (event.mKeyCode) {
-                    KeyEvent.KEYCODE_F1 -> {
-
-                    }
-                    KeyEvent.KEYCODE_F2 -> {
-                        Log.d(TAG, "Redeploying...")
-                        coroScope.launch { rime.deploy() }
-                        return
-                    }
-                }
                 sendDownUpKeyEvent(event.mKeyCode, event.mX)
             }
 
@@ -922,16 +506,7 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
     override fun onCustomRequest(requestCode: Int): Boolean = false
     override fun onMovingCursorLockEvent(canMoveCursor: Boolean) {}
 
-    override fun clearUserHistoryDictionaries() {
-        Log.d(TAG, "Clearing Chinese dictionary...")
-
-        coroScope.launch {
-            rime.shutdown()
-            getUser(helper.context).deleteRecursively()
-            rime.startup(false)
-            prevConfiguration = null
-        }
-    }
+    override fun clearUserHistoryDictionaries() {}
 
     override fun onStartBatchInput() {}
     override fun onUpdateBatchInput(batchPointers: InputPointers?) {}
@@ -947,5 +522,5 @@ class ChineseIME(val helper: IMEHelper) : IMEInterface, SuggestionStripViewAcces
 // Non-behavior methods }}
 
     val debugInfo: String
-        get() = "configuration=${prevConfiguration}\nlayoutHint=${layoutHint}\nlocale=${Settings.getInstance().current.mLocale}\nisSimplified=${isSimplifiedChinese(Settings.getInstance().current.mLocale)}\nrawInput=${rawInput.text}"
+        get() = "layoutHint=${layoutHint}\nlocale=${Settings.getInstance().current.mLocale}\nrawInput=${rawInput.text}"
 }
